Preview bars update, value logged to list
Update speed is, in fact, not sufficient.
